add licence for using the library of module_advanced_foc
module_motorcontrol_licence is added to sc_sncn_motorcontrol_new repository. This new module contains a constant value which is passed to Motor_Control_service() task, and will be checked each time the application starts.
If this constant is not matching, then Motor_Control_Service() task will give a "wrong licence" error, and will not start to work.
Off course, this new module should also be imported in the project. Later on, this new module will be replaced by a licence file which should be sent for the end user of the software