motion control service: improve initialization and update of brake parameters, also implement brake disable strategy.
Now the brake configuration is initialized when the service starts.
The brake config is non blocking but it stops the motorcontrol and will delay the next start for 2 seconds.
(if a client wants to start it when the brake config is not finished)
The brake config is also automatically started when the brake parameters are changed.
(again it will stop the control so the client have to take care of that)

Brake release strategy:
now 0 will disable the brake
1 is normal release
2 to 100 is shaking release with a percentage of max torque.