update the error_sens member of upstream_data_struct with the result of auto offset detection routine.
so: imagine there is no error. Now, if an error is detected in position_feedback_service, then this error will be saved in error_sens member of upstream_data_struct.
However, if under this faulty condition the user triggers the auto offset detection, then this error will be overwritten with the error which is found by offset detection routine.
if the user sends reset offset command, all error statuses will be cleared.